reestructure entire project with MVC structure, separate enum class and its test and rename branches, classes and packages
