--- src/main/java/dev/erica/ghostbusters/Ghost/CatchGhost.java
@@ -1,5 +1,0 @@
-package dev.erica.ghostbusters.Ghost;
-
-public class CatchGhost {
-
-}
