+++ src/main/java/dev/erica/ghostbusters/controller/GhostController.java
@@ -1,0 +1,5 @@
+package dev.erica.ghostbusters.controller;
+
+public class GhostController {
+
+}
