+++ src/main/java/dev/erica/ghostbusters/controller/UserController.java
@@ -1,0 +1,5 @@
+package dev.erica.ghostbusters.controller;
+
+public class UserController {
+
+}
