+++ src/main/java/dev/erica/ghostbusters/model/GhostModel.java
@@ -1,29 +1,6 @@
-package dev.erica.ghostbusters.Ghost;
+package dev.erica.ghostbusters.model;
 
-public class Ghost {
-
-    
-
-    public enum GhostClass {
-        CLASS_I("Clase I - Manifestación menor"),
-        CLASS_II("Clase II - Aparición móvil"),
-        CLASS_III("Clase III - Entidad inteligente"),
-        CLASS_IV("Clase IV - Fantasma histórico"),
-        CLASS_V("Clase V - Espíritu antropomorfo"),
-        CLASS_VI("Clase VI - Espíritu demoníaco"),
-        CLASS_VII("Clase VII - Entidad ultraterrena");
-
-        private final String description;
-
-        GhostClass(String description) {
-            this.description = description;
-        }
-
-        public String getDescription() {
-            return description;
-        }
-
-    }
+public class GhostModel {
 
     private static int idCounter = 1;
     private final int id;
@@ -33,7 +10,7 @@ public class Ghost {
     private final String ability;
     private final String captureDate;
 
-    public Ghost(String name, GhostClass ghostClass, String dangerLvl, String ability, String captureDate) {
+    public GhostModel(String name, GhostClass ghostClass, String dangerLvl, String ability, String captureDate) {
 
         this.id = idCounter++;
         this.name = name;
