+++ src/main/java/dev/erica/ghostbusters/model/UserModel.java
@@ -1,0 +1,5 @@
+package dev.erica.ghostbusters.model;
+
+public class UserModel {
+
+}
