+++ src/main/java/dev/erica/ghostbusters/view/GhostView.java
@@ -1,0 +1,5 @@
+package dev.erica.ghostbusters.view;
+
+public class GhostView {
+
+}
