--- src/test/java/dev/erica/ghostbusters/Ghost/CatchGhostTest.java
@@ -1,28 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/UnitTests/JUnit4TestClass.java to edit this template
- */
-
-package dev.erica.ghostbusters.Ghost;
-
-import static org.junit.Assert.fail;
-import org.junit.Test;
-import org.junit.jupiter.api.DisplayName;
-
-/**
- *
- * @author Erica
- */
-public class CatchGhostTest {
-
-    public CatchGhostTest() {
-    }
-
-    @Test
-    @DisplayName("")
-    public void testSomeMethod() {
-        // TODO review the generated test code and remove the default call to fail.
-        fail("The test case is a prototype.");
-    }
-
-}
