+++ src/test/java/dev/erica/ghostbusters/controller/UserControllerTest.java
@@ -1,0 +1,4 @@
+package dev.erica.ghostbusters.controller;
+
+public class UserControllerTest {
+}
