+++ src/test/java/dev/erica/ghostbusters/model/GhostModelTest.java
@@ -1,27 +1,22 @@
-package dev.erica.ghostbusters.Ghost;
+package dev.erica.ghostbusters.model;
 
 import org.junit.Test;
 import org.junit.jupiter.api.DisplayName;
+
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.Matchers.greaterThan;
 
-public class GhostTest {
 
-    @Test
-    @DisplayName("Test GhostClass constructor")
-    public void testGhostClass() {
-        Ghost.GhostClass ghostClass = Ghost.GhostClass.CLASS_I;
-        assertThat(ghostClass.getDescription(), is("Clase I - Manifestación menor"));
-    }
+public class GhostModelTest {
 
     @Test
     @DisplayName("Test GhostConstructor")
     public void testGhostConstructor() {
-        Ghost ghost = new Ghost("Marcelino Paxarada", Ghost.GhostClass.CLASS_II, "Bajo", "Leer la mente", "12-02-2025");
+        GhostModel ghost = new GhostModel("Marcelino Paxarada", GhostClass.CLASS_II, "Bajo", "Leer la mente", "12-02-2025");
         assertThat(ghost.getID(), greaterThan(0));
         assertThat(ghost.getName(), is("Marcelino Paxarada"));
-        assertThat(ghost.getGhostClass(), is(Ghost.GhostClass.CLASS_II));
+        assertThat(ghost.getGhostClass(), is(GhostClass.CLASS_II));
         assertThat(ghost.getDangerLvl(), is("Bajo"));
         assertThat(ghost.getAbility(), is("Leer la mente"));
         assertThat(ghost.getCaptureDate(), is("12-02-2025"));
