+++ src/test/java/dev/erica/ghostbusters/model/UserModelTest.java
@@ -1,0 +1,4 @@
+package dev.erica.ghostbusters.model;
+
+public class UserModelTest {
+}
