+++ src/test/java/dev/erica/ghostbusters/view/GhostViewTest.java
@@ -1,0 +1,4 @@
+package dev.erica.ghostbusters.view;
+
+public class GhostViewTest {
+}
