+++ src/test/java/dev/erica/ghostbusters/view/UserViewTest.java
@@ -1,0 +1,4 @@
+package dev.erica.ghostbusters.view;
+
+public class UserViewTest {
+}
